+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/updates/queues/AnonymityRevokerInfo.java
@@ -1,11 +1,16 @@
 package com.concordium.sdk.responses.blocksummary.updates.queues;
 
+import com.concordium.sdk.crypto.elgamal.ElgamalPublicKey;
+import com.concordium.sdk.serializing.JsonMapper;
 import com.concordium.sdk.types.UInt32;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.google.common.collect.ImmutableList;
+import concordium.ConcordiumP2PRpc;
+import lombok.*;
+
+import java.util.Optional;
 
 /**
  * Anonymity revoker info
@@ -24,13 +29,28 @@ public final class AnonymityRevokerInfo {
      */
     private final Description description;
 
-    private final String anonymityRevokerPublicKey;
+    /**
+     * Elgamal encryption key of the anonymity revoker.
+     */
+    private final ElgamalPublicKey anonymityRevokerPublicKey;
 
     @JsonCreator
-    AnonymityRevokerInfo(@JsonProperty("arIdentity") int arIdentity, @JsonProperty("arDescription") Description description, @JsonProperty("arPublicKey") String arPublicKey) {
+    public AnonymityRevokerInfo(
+            @JsonProperty("arIdentity") int arIdentity,
+            @JsonProperty("arDescription") Description description,
+            @JsonProperty("arPublicKey") String arPublicKey) {
         this.arIdentity = UInt32.from(arIdentity);
         this.description = description;
-        this.anonymityRevokerPublicKey = arPublicKey;
+        this.anonymityRevokerPublicKey = ElgamalPublicKey.from(arPublicKey);
     }
 
+    public static Optional<ImmutableList<AnonymityRevokerInfo>> fromJsonArray(ConcordiumP2PRpc.JsonResponse res) {
+        try {
+            val array = JsonMapper.INSTANCE.readValue(res.getValue(), AnonymityRevokerInfo[].class);
+
+            return Optional.ofNullable(array).map(arr -> ImmutableList.copyOf(arr));
+        } catch (JsonProcessingException e) {
+            throw new IllegalArgumentException("Cannot parse Anonymity Revoker JSON", e);
+        }
+    }
 }